Routes inserted
Tweets and tweet stats plus unit tests are inserted into the routes plus both of their blueprints are registered in the app
